refactor: minor improvements and refactoring
