+++ components/HomePage/Kanban/Column/JobPosts/JobModal/JobDocuments/Documents.tsx
@@ -18,7 +18,6 @@ import {
   attachDocumentToJobApplication,
   detachDocumentFromJobApplication,
 } from '@/redux/documents/documentsThunk';
-// Removed unused import
 import EditDocumentModal from '@/components/Forms/AddDocument/EditDocumentModal';
 
 const Documents = () => {
@@ -27,9 +26,10 @@ const Documents = () => {
   const jobs = useAppSelector((state) => state.jobs);
   const user = useAppSelector((state) => state.user);
   const userDocuments = useAppSelector(selectUserDocuments);
-  // We don't need isUploadModalOpen state as the UploadDocumentModal handles its own visibility
   const [isEditModalOpen, setIsEditModalOpen] = useState(false);
-  const [documentToEdit, setDocumentToEdit] = useState<JobDocument | null>(null);
+  const [documentToEdit, setDocumentToEdit] = useState<JobDocument | null>(
+    null
+  );
   const accessToken = (() => {
     try {
       return localStorage.getItem('accessToken');
